% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{bootstrap_se}
\alias{bootstrap_se}
\title{The bootstrap se, based on Algorithm 2 of Arkhangelsky et al.}
\usage{
bootstrap_se(estimate, replications)
}
\arguments{
\item{estimate}{multi_sdid object}

\item{replications}{number of replicates}
}
\description{
The bootstrap se, based on Algorithm 2 of Arkhangelsky et al.
}
