% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collapsed.form}
\alias{collapsed.form}
\title{Collapse Y to an N0+1 x T0+1 x J vector by averaging the last N1=nrow(Y)-N0 rows and T1=ncol(Y)-T0 columns
it is the modified version of Arkhangelsky et al for}
\usage{
collapsed.form(Y, N0, T0)
}
\description{
Collapse Y to an N0+1 x T0+1 x J vector by averaging the last N1=nrow(Y)-N0 rows and T1=ncol(Y)-T0 columns
it is the modified version of Arkhangelsky et al for
}
