% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{german_reunification}
\alias{german_reunification}
\title{German reunification on West Germany}
\format{
To be updated
A data frame with:
\describe{
  \item{country}{Country name}
  \item{gdp}{GDP}
  \item{year}{Year}
  ...
}
}
\usage{
data(german_reunification)
}
\description{
German reunification on West Germany
}
\examples{
\donttest{
# Load German reunification long panel format
data("german_reunification")
}

}
