% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_sdid_estim.R
\name{multi_synthdid_curves}
\alias{multi_synthdid_curves}
\title{Outputs the synthetic, treated units and the
effect curve that was averaged to produce our point estimate
 (at the moment it may give slightly different results
if standardization has been used during the optimization )}
\usage{
multi_synthdid_curves(estimate, complete = F)
}
\arguments{
\item{estimate, }{as output by multi_synthdid_estim}

\item{complete, }{TRUE/FALSE.
- TRUE: it contains pre- and post event periods as well.
- FALSE: it only contains after event values (Default)}
}
\value{
list with:
       - synt_curve: synthetic outcome
       - tr_curve: treated (averaged) unit outcome
       - tau_curve: effect/gap curve
}
\description{
Outputs the synthetic, treated units and the
effect curve that was averaged to produce our point estimate
 (at the moment it may give slightly different results
if standardization has been used during the optimization )
}
