% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_sdid_estim.R
\name{multi_synthdid_placebo}
\alias{multi_synthdid_placebo}
\title{Computes a placebo variant of our estimator using pre-treatment data only}
\usage{
multi_synthdid_placebo(estimate, treated.fraction = NULL)
}
\arguments{
\item{estimate, }{as output by multi_synthdid_estim}

\item{treated.fraction, }{the fraction of pre-treatment data to use as a placebo treatment period
Defaults to NULL, which indicates that it should be the fraction of post-treatment to pre-treatment data}
}
\description{
Computes a placebo variant of our estimator using pre-treatment data only
}
