% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{placebo_se}
\alias{placebo_se}
\title{The placebo se, based on Algorithm 4 of Arkhangelsky et al.}
\usage{
placebo_se(estimate, replications)
}
\arguments{
\item{estimate}{multi_sdid object}

\item{replications}{number of replicates}
}
\description{
The placebo se, based on Algorithm 4 of Arkhangelsky et al.
}
