% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msdid_plots.R
\name{plot_gaps}
\alias{plot_gaps}
\title{Plotting functions for multi_sdid objects}
\usage{
plot_gaps(
  estimate,
  outcome = NULL,
  placebo = F,
  se_type = "placebo",
  se_replication = 150,
  max_rep_plot = 150
)
}
\arguments{
\item{estimate}{milti_sdid object}

\item{outcome}{char vector of outcomes to plot}

\item{placebo}{TRUE/FALSE to plot the placebo outcomes along with the gaps}

\item{se_type}{`placebo` or `bootrstrap`}

\item{se_replication}{how many time SE should be calculated}

\item{max_rep_plot}{how many replication to plot}
}
\description{
Plotting functions for multi_sdid objects
}
