% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msdid_plots.R
\name{plot_outcomes}
\alias{plot_outcomes}
\title{Plotting functions for multi_sdid objects}
\usage{
plot_outcomes(estimate, outcome = NULL)
}
\arguments{
\item{estimate}{milti_sdid object}

\item{outcome}{char vector of outcomes to plot}
}
\description{
Plotting functions for multi_sdid objects
}
