% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.multisynthdid}
\alias{predict.multisynthdid}
\title{Predict a multisynthdid object}
\usage{
\method{predict}{multisynthdid}(
  object,
  formula_yx,
  treatment,
  unit,
  time,
  newdata,
  SE = T,
  se_type = "placebo",
  se_replication = 200
)
}
\arguments{
\item{object}{The object to predict}

\item{formula_yx}{formula containing the variables/structure to predict}

\item{newdata}{the new data to use for prediction}
}
\description{
Predict a multisynthdid object
}
