% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{sc.weight.fw}
\alias{sc.weight.fw}
\title{A Frank-Wolfe solver for synthetic control weights using exact line search}
\usage{
sc.weight.fw(
  Y,
  zeta,
  theta,
  intercept = TRUE,
  weight = NULL,
  min.decrease = 0.001,
  max.iter = 1000,
  scale_vcov = T,
  smooth = F,
  time_FE = F
)
}
\description{
A Frank-Wolfe solver for synthetic control weights using exact line search
}
