% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{se.multi_sdid}
\alias{se.multi_sdid}
\title{Calculate Variance-Covariance Matrix for a Fitted Model Object}
\usage{
\method{se}{multi_sdid}(object, method = c("bootstrap", "placebo"), replications = 200)
}
\arguments{
\item{object}{A synthdid model}

\item{method, }{the CI method. The default is placebo (warning: this may be slow on large
data sets).}

\item{replications, }{the number of bootstrap replications}
}
\description{
Provides variance estimates based on the following two options following Arkangelsky et al.
\itemize{
  \item The bootstrap, Algorithm 2 in Arkhangelsky et al.
  \item Placebo, Algorithm 4 in Arkhangelsky et al.
  \item Jacknife is not applicable
}
}
\details{
"placebo" is the only option that works for only one treated unit.
}
\references{
Agoston Reguly, Chava Sudheer.
 "The Long-Run Stock Market Performance of Mergers and Acquisitions". SSRN prepint
 https://papers.ssrn.com/sol3/papers.cfm?abstract_id=4742401

Dmitry Arkhangelsky, Susan Athey, David A. Hirshberg, Guido W. Imbens, and Stefan Wager.
 "Synthetic Difference in Differences". arXiv preprint arXiv:1812.09970, 2019.
}
