% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup_multi_sdid}
\alias{setup_multi_sdid}
\title{Convert a long (balanced) panel to a wide multi dimensional matrices (tensors)}
\usage{
setup_multi_sdid(
  df,
  unit,
  time,
  outcome,
  treatment,
  covariates = NULL,
  eq_s = NULL
)
}
\arguments{
\item{df}{A data.frame with columns consisting of units, time, outcome, and treatment indicator.}

\item{unit}{The column name corresponding to the unit identifier.}

\item{time}{The column name corresponding to the time identifier.}

\item{outcome}{The column names corresponding to the outcome identifiers (can be multiple -- 1xJ).}

\item{treatment}{The column number corresponding to the treatment status.}

\item{covariates}{name of the used covariates from the data frame. Default is no covariates.}

\item{eq_s}{character matrix with LxK elements. J is in the same order as `outcome` is given.
The K elements may be blank or containing the covariates referring to the j-th outcome equation.}
}
\value{
A list with entries `Y`: the data tensor, `N0`: the number of control units, `T0`:
 the number of time periods before treatment, `W`: the matrix of treatment indicators,
 `X` the data tensor for covariates and `unit`: standing for the ordering of control + treated units.
}
\description{
Converts a data set in panel form to matrix format required by multi_synthdid estimators.
A typical long panel date set looks like \[unit, time, outcomes, treatment, covariates\].
Multi dimensional version of synthdid requires a balanced panel with simultaneous
adoption of treatment: each unit must be observed at all times, and all treated units
must begin treatment simultaneosly. This function
creates num.units x num.time.periods x num.outcome.vars matrices Y and
num.units x num.time.periods for treatment matrix W. It also creates a covariance matrix
X with 4 dimensions: num.units x num.time.periods x outcome_var x selected.covariates
The 3rd dimension specify for which outcome the covariate is used and the 4th specify the
selected covariate.
In these matrices, columns are sorted by time, and rows for control units
appear before those of treated units.
}
\examples{
\donttest{
# Load tobacco sales in long panel format.
data("german_reunification")
# Transform to N*T matrix format required for multi_synthdid_estim,
# where N is the number of units and T the time periods and
# J is the used number of outcome variables.
setup <- setup_multi_sdid(german_reunification, unit = 'country',
                          time = 'year', outcome = c('gdp','infrate','trade'),
                          treatment = 'W')

# Compute synthdid estimate
multi_synthdid_estim(setup$Y, setup$N0, setup$T0)
}

}
