% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_sdid_estim.R
\name{sparsify_function}
\alias{sparsify_function}
\title{A function mapping a numeric vector to a (presumably sparser) numeric vector of the same shape to
be passed onto multi_synthdid_estim}
\usage{
sparsify_function(v)
}
\arguments{
\item{v}{a vector}
}
\description{
A function mapping a numeric vector to a (presumably sparser) numeric vector of the same shape to
be passed onto multi_synthdid_estim
}
