% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.multi_sdid}
\alias{summary.multi_sdid}
\title{Summarize a synthdid object}
\usage{
\method{summary}{multi_sdid}(object, weights = F, weight.digits = 3, weights_print = "non-null")
}
\arguments{
\item{object}{The object to summarize}

\item{weights}{Report weights or not}

\item{weight.digits}{The number of digits to use when displaying weights (omega, lambda)}

\item{weights_print}{`non-null` - only report non-null values, `all` - report all values.}
}
\description{
Summarize a synthdid object
}
