% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.multisynthdid}
\alias{summary.multisynthdid}
\title{Summarize a multisynthdid object}
\usage{
\method{summary}{multisynthdid}(object, weights = F, weight.digits = 3, weights_print = "non-null")
}
\arguments{
\item{object}{The object to summarize}

\item{weights}{Report weights or not}

\item{weights_print}{`non-null` - only report non-null values, `all` - report all values.}
}
\description{
Summarize a multisynthdid object
}
