% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{german_reunification}
\alias{german_reunification}
\title{German reunification on West Germany}
\format{
## `german_reunification`
A data frame with:
\describe{
  \item{country}{Country name}
  \item{gdp}{GDP}
  \item{year}{Year}
  ...
}
}
\usage{
german_reunification
}
\description{
German reunification on West Germany
}
\keyword{datasets}
